flashbench: add initial runtime unit
This is the start of a rewrite of the entire program,
based on the idea of defining test cases dynamically
in pseudo-code. A test case is written as a C data
structure containing basic operations (read, write,
erase, ...), output (text, times, ...) and loops
to repeat operations with different positions and
sizes.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>