flashbench: add --offset command line argument
This makes it easier to do the --open-au test case
on drives whose erase block is not a power of two.
Note that for those, you have to pass a blocksize
that is also not a power of two, but a fraction
of the erase size, e.g.

flashbench --open-au --random --open-au-nr=6 \
	--erasesize=$[3 * 1024 * 1024] \
	--blocksize=$[24 * 1024] \
	--offset=$[24 * 1024 * 1024] \
	/dev/sde

Signed-off-by: Arnd Bergmann <arnd@arndb.de>